Default forwarded header handler should only using ConnectionInfo public API
